implement feature to allow load balancing on the hardware behind resources
The ideas is that often several logical resources actually shared hardware in
he background and there is a desire to spread the build load evenly among the
real physical hardware and not just the logical resources.

Define one or more "load balancing" labels.  Resources that share a label
which has been listed a "load balancing" will be locked in a fashion so as to
minimize the load on each "load balanced" label.