refactoring to add form validation to resource manager page
The previous refactoring to pull the resource manager configuration into its
own page under the system management page left things a little messy and it was
difficult, if not impossible to get live form validation working.  This change
refactors LockableResourcesManager as a legacy plugin extension and implements
basic form validation.