revamp queue/locking logic
This revamp was necessary in order to correctly support the new required
resources parameter type.  The previous locking mechanism was too focused on
storing the list of required resources in the job config.

The new architecture uses the LockedResourcesBuildAction as a data structure to
keep track of which resources are locked for a particular build instance.  This
has the advantage of having Jenkins pass the data structure around for us,
including once the build actually starts (when the queueID is no longer available)
instead of having to try to keep track of things using a combination of the job
name and queueID.