avoid locking resources with parent builds of matrix projects
The problem was that we have always locked the resources. However, this
makes no sense for parent build of matrix projects since they dont
actually do anything. To fix the problem we just ignore matrix builds
(that do nothing but spawn configuration builds). Now we consume exactly
the same number of resource instances as actually needed.

Signed-off-by: Nickolay Martinov
Signed-off-by: Robin Jarry <robin.jarry@6wind.com>