allow an internal system script to alter a resource's labels on the fly
The goal is to make it easy for an authorized admin level script to
programmatically alter the labels attached to a resource without having to
work through the UI specific configure method. This allows for resources which
are manged by other Jenkins jobs to be updated with appropriate labels as their
state changes.