Added functionality for reserving resources to specific nodes
Reserving resources to nodes:
- allows certain resources to be locked only by nodes that can access them
- allows finer control over which node can take hold of resources at any point

This is a simple feature that is added to the resources configuration as another field.
Functionality is very similar to reserving resources for users, however, a resource can be
reserved for multiple nodes at any given time (any of the nodes in the list can use the resource
and lock or queue it). Allowing multiple nodes to lock one resource is useful if
the test grid is large and not all test nodes have direct access to the resource, and
attempting to use it would result in errors or delays.

NOTE: In this context, the "nodes" are the slaves, and Jenkins cannot tell the difference
between two slaves running on the same machine, or two different machines.

Changelog (for reserving resources to nodes):
LockableResource.java
-> added reservedForNode and reservedForNodes fields
-> added get/set methods for the above fields
-> added unReserveForAll method that can free the node of all reservations (for nodes)

LockableResourcesManager.java
-> added node parameter to the findAvailableResources used to verify if the node that requests the resources can have them
-> added node parameter to queueId method used to check if the resources can be used by the given node
-> added conditions to check if the resources can be queued/locked to the following methods:
	- findAvailableResources
	- queueId
	- lock
	- reserve

LockableResouresQueueTaskDispatcher.java
-> added new BecauseWrongNode as a viable cause of blockage that can be returned by canTake

LockableResourceTest.java
-> added test for getReservedForNode, isReservedForNode and setReservedForNode methods (similar to the
tests for reservedBy)

Utils.java
-> added getNodeName() method that returns the name of the node that runs this method, based on the
current executor

LockableResource.jelly
-> added new field for reserved for node
-> changed description for 'reservedBy' to 'reservedByUser'

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>