Added initializer method to update the reservedForNodesSets at startup
This change fixes loading resources from the hard drive at jenkins
startup. The problem was caused by having the set of node names as
transient, thus not being saved on the drive. When jenkins tries to
find that information, the reading fails, and the set is null.
Now, during startup the list is re evaluated.

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>