Prepare for resource automatic generation on system config change
Changed Jenkins main version to 1.609.2, and added dependency for
the matrix-project plugin. This update was required in order to correctly
receive notifications regarding node change for the master node. The trigger
for master node configuration change was added in 1.571, and Jenkins main
version might be changed to that without causing any structural errors.
Should the decision to rollback to that version be taken, the matrix-project
plugin would still be required, as it has been separated from the main branch
since Jenkins version 1.561.

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>