Added functionality for breaking automatically generated resources
Resource breaking is used when an attempt to remove automatically
generated resources is made (using LockableResourcesManager.
removeAutoGenResourcesForNode), but the resource was previously
locked by a build. Broken resources are removed during unlock.

Changelog:
LockableResources.java
-> added a new 'isBroken' field that is true if the resource was broken.
-> added 'getIsBroken()' method for the 'isBroken' field.
-> added 'breakResource()' method that sets the 'isBroken' field to true
and changes the description to a string explaining the current state of
the resource.

LockableResourcesManager.java
-> added 'removeBrokenResource(LockableResource) method that removes the
given resource if it is broken.
-> refactored the 'unlock' method to include a call to
'removeBrokenResource'.
-> refactored 'removeAutoGenResourcesForNode' to break the resource if it
is in use, instead of removing it immediately.

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>