Refactored matrix configuration computation method
Refactored 'DynamicUtils.getDynamicResConfig' method; the method
now correctly uses the matrix configuration of the project,
instead of splitting the job name in order to obtain it.

Added a method to retrieve the matrix configuration of the project a
build is part of in DynamicUtils.java named 'getProjectConfiguration'.

Also, changed the type for 'ignoredAxis' in 'getDynamicResConfig' from
List to Set to lower the time required to find names.

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>