Removed the injectedId field
The program no longer requires manually defining injectedIds and values,
to be able to differentiate between runs.
Each time a master job ends, it creates tokens for downstream projects that use
dynamic resources. These tokens are identified in a map using the name and
build number of master jobs.

Changelog:
dynamicres.DynamicInfoData.java
-> renamed 'injectedValue' to 'tokenValue'
-> the value of the tokens is now retrieved using 'DynamicUtils.
getProjectToken'

dynamicres.DynamicResourcesProperty.java
-> removed 'injectedId' field
-> removed get method for 'injectedId'
-> removed 'doCheckInjectedId' method
-> refactored 'newInstance' method to no longer define 'injectedId'

dynamicres.DynamicUtils.java
-> added 'DYNAMICRES_TOKENID' as a string used to identify the token when
viewing dynamic resources configurations
-> added 'tokens' as a map used to store tokens created for projects
-> added 'getCurrentRunTokenID' and 'getFutureRunTokenID' methods to get a
key to identify tokens for projects. The CurrentRun method is used to
create an id for a project already running, while the FutureRun method is
used to create an id for a project that is not yet running
-> refactored the 'getDynamicResConfig' to use tokens instead of injected
parameters
-> removed 'getBuildVariables' and 'getBuildVariableValue' methods that
were no longer required
-> added 'getProjectToken' methods that will retrieve the value of the
token created for a project. If no value is found, one will be created
based on the current time
-> added 'createTokensForDownstream' that will create tokens for all
downstream projects that use dynamic resources, based on the token
created for the build received as parameter. If no token is found, one will
be created based on the current time
-> refactored 'buildEnded' method to also remove the token created for the
project that the given build is part of

LockRunListener.java
-> 'onCompleted' method will now create tokens for downstream projects if
a MatrixBuild has completed

DynamicResourcesProperty - config.jelly
-> removed 'injectedId' field

help-injectedId.html - removed

help-consumeDynamicResources.html
help-createDynamicResources.html
-> changed information to specify tokens instead of injected parameters

--> reworked documentation to accomodate new feature

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>