Removed call to remove tokens in 'DynamicUtils.buildEnded'
The resource that the program attempted to remove based on the 'unique
name' does not exist. Tokens are created based on master job name alone
(the unique name also includes matrix configuration).
Tokens are removed by 'createTokensForDownstream' after tokens for
downstream jobs are created and should not be removed earlier.

Signed-off-by: Darius Mihai <darius.mihai@freescale.com>