[FIXED JENKINS-40879] Only call proceed with resources to lock.
Previously, we were calling LockStepExecution.proceed with the full
list of resources - which was resulting in *everything* getting
unlocked if there was a next context, i.e., if anything else was
waiting on a resource to be unlocked. That was wrong, obviously.

This changes to instead only call proceed with the list of resource
names we actually need to lock for the next context, which fixes the
test in the previous commit.