This adds the `tryLock` and `releaseLock` steps. It also enables using `lock` without a body.
The `tryLock` can be used like this:

```groovy
if (tryLock(resource: 'staging-server')) {
  try {
    println "Resource was acquired :)"
  } finally {
    releaseLock(resource: 'staging-server')
  }
} else {
    println "Resource was not acquired :("
}
```

fixes #164 and JENKINS-43336

= Searching for available resource

Sometimes we have many resources, perhaps 1000. We have no need for the queue feature in this plugin, we only need to make sure we have exclusive access to a resource. We know there are free resources, it is just a matter of finding one. We may need to find a port to use when starting a server. Then we can do:

```groovy
node {
  // Find a free resource
  def port = -1
  stage("Find port") {
    for (int i = 0; i < 1000; i++) {
      port = i
      if (tryLock(resource: 'staging-server-' + i)) {
        break;
      }
    }
  }

  stage("Use port") {
    try {
      // Start server on port
    } finally {
      releaseLock()
    }
  }
}
```

Also one can implement an `isAvailable` method by doing:

```groovy
boolean isAvailable(resource) {
  def isAvailable = tryLock(resource: resource)
  if (isAvailable) {
    releaseLock(resource: resource)
  }
  return isAvailable
}
```

= Visualize the locking in its own stage

By not requiering `lock` to have a body, we can put the `lock`-part in a `stage` and explicitly release it later. Visualizing how much time we spend on waiting for an available resource.

```groovy
node {
  stage("Waiting for resource") {
    lock(resource: 'myResource')
  }

  stage("Using resource") {
    try {
      println "Using the resource here..."
    } finally {
      println "Now releasing the resource..."
      releaseLock()
    }
  }
}
```