Don't remove non-locked resources
As far as I can tell, the intention to reset non-locked resources to
default, is handled in setDeclaredResources. With moving the
copyUnconfigurableProperties to setDeclaredResources, this line actually
causes problems, because it removes current resources from the list, but
we need to keep them, so we can copy e.g. the note/reservations from
them. Note is not shown in the UI, so it's also not part of the data
model supplied to setDeclaredResources. When keeping this line, it
causes the note to disappear when saving resources in UI configuration.