Ignore reservedBy/note at Jenkins initalisation with CasC
This makes sure that when a clean Jenkins instance starts with CasC, the
reservedBy/note are not loaded from CasC if they are defined. Or when
loading CasC with new resources.

This should not affect UI configuration. reservedBy/note fields are not
shown, so they should always be empty for new resources.