Ignore unconfigurable properties in casc config (#771)
* Keep existing reservations/notes when reloading casc

Only when they are not defined explicitly in casc configuration

* Ignore unconfigurable properties from CasC configuration

This includes reservedBy, reservedTimestamp and note.

Also remove the reservedBy from the UI configuration.

* Don't remove non-locked resources

After the previous changes to ignore unconfigurable properties, this
causes all reservations to disappear. As far as I can tell, the
intention to reset non-locked resources to default, is handled in
setDeclaredResources.

* Check that reservedBy/note are empty in casc tests

Also include non-configurable properties in CasC configuration tests, to
make sure there are no errors when loading it.

* Ignore reservedBy/note at Jenkins initalisation with CasC

This makes sure that when a clean Jenkins instance starts with CasC, the
reservedBy/note are not loaded from CasC if they are defined. Or when
loading CasC with new resources.

This should not affect UI configuration. reservedBy/note fields are not
shown, so they should always be empty for new resources.

---------

Co-authored-by: Benjamin Meeusen <benjamin.meeusen@septentrio.com>
Co-authored-by: Martin Pokorny <89339813+mPokornyETM@users.noreply.github.com>